Wayland clipboard support via data_control protocol
While it is true that wayland can support that X11 way of handling
clipboard support, this requires a running xwayland server and
compositor support for running xwayland.

This change introduces a new clipboard context
`WaylandDataControlClipboardContext` which is a wrapper around the
wl-clipboard-rs crate. This crate handles all of the communication with
the compositor for us.

To selected between X11 and wayland data_control clipboard handle is not
possible via a platform selector, as is the case for windows and mac,
two new features where add `x11` and `wayland-data-control`. The `x11`
feature is in the `default` feature set as to preserve backward
compatibility.

The feature name `wayland-data-control` was chosen to handle other
possible wayland protocols to interface with a clipboard.