Remove macOS memory leaks by properly passing objects via msg_send
Previously, a number of functions passed some runtime-owned object by
value, in Rust terms. However, the system pasteboard APIs were only
expecting a reference (pointer) to the object instead (like many
Objective-C APIs do). As a result, the object was moved into the FFI
which was, as expected, not attempting to free the object when it was
done.