Remove leaks from macOS clipboard getters
This additionally refactors `image()` to be simpler by taking advantage
of the fact that we only want one image. Additionally, this codepath avoids
the system-level caching of image data that can help reduce peak usage.
We have applied an autorelease pool to the Obj-C code as well, as there
appear to be a lot of autoreleased values used by AppKit.framework

Finally, it applies a similar refactoring to `text()` for simplicity and
removal of memory leaks due to incorrect object retaining assumptions.

Co-authored-by: taojiacheng <swkzymlyy@gmail.com>